Almost finished

+++ adapter/image.cpp
@@ -15,5 +15,5 @@ bool operator<(const signedPixel& left, const signedPixel& right) {
     return temp1 < temp2;
 }
 signedPixel operator+(const signedPixel& left, const signedPixel& right) {
-    return signedPixel{left.red + right.red, left.green + right.green, left.blue + right.blue};
+    return signedPixel{(int16_t)(left.red + right.red), (int16_t)(left.green + right.green), (int16_t)(left.blue + right.blue)};
 }
